+++ src/knot.erl
@@ -1,4 +1,4 @@
--module(rotor).
+-module(knot).
 
 -on_load(init/0).
 
@@ -6,9 +6,9 @@
 
 -spec init() -> ok.
 init() ->
-    SoName = filename:join(code:priv_dir(rotor), "rotor"),
+    SoName = filename:join(code:priv_dir(knot), "knot"),
     erlang:load_nif(SoName, 0).
 
 -spec uniform(pos_integer()) -> pos_integer().
 uniform(_N) ->
-    erlang:nif_error(rotor_nif_not_loaded).
+    erlang:nif_error(knot_nif_not_loaded).